Update Packaging
Moving Package Model Inside The Main Package (DTO)
