+++ BE/src/main/java/com/example/dto/model/LevelModel.java
@@ -1,4 +1,4 @@
-package com.example.model;
+package com.example.dto.model;
 
 
 import javax.persistence.Column;
